pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ugc-ai-influencer-bot"
version = "0.1.0"
description = "Telegram-first MVP for generating UGC-style AI influencer video packages."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
